+++ targets/TARGET_STM/TARGET_STM32F0/TARGET_NUCLEO_F072RB/objects.h
@@ -66,10 +66,6 @@ struct dac_s {
     uint32_t channel;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-};
-
 struct can_s {
     CANName can;
     int index;
