+++ targets/TARGET_STM/TARGET_STM32F3/TARGET_DISCO_F334C8/objects.h
@@ -66,11 +66,6 @@ struct dac_s {
     uint32_t channel;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-    uint32_t slave;
-};
-
 #include "common_objects.h"
 #include "gpio_object.h"
