+++ targets/TARGET_STM/TARGET_STM32F7/TARGET_DISCO_F746NG/objects.h
@@ -66,11 +66,6 @@ struct dac_s {
     uint32_t channel;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-    uint32_t slave;
-};
-
 struct can_s {
     CANName can;
     int index;
