+++ targets/TARGET_STM/TARGET_STM32L0/TARGET_NUCLEO_L031K6/objects.h
@@ -60,10 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-};
-
 #include "common_objects.h"
 #include "gpio_object.h"
