+++ targets/TARGET_STM/TARGET_STM32L0/TARGET_NUCLEO_L073RZ/objects.h
@@ -66,10 +66,6 @@ struct dac_s {
     uint32_t channel;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-};
-
 struct trng_s {
     RNG_HandleTypeDef handle;
 };
