+++ targets/TARGET_STM/TARGET_STM32L1/TARGET_NUCLEO_L152RE/objects.h
@@ -65,11 +65,6 @@ struct dac_s {
     PinName pin;
 };
 
-struct i2c_s {
-    I2CName  i2c;
-    uint32_t slave;
-};
-
 #include "common_objects.h"
 #include "gpio_object.h"
